pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-dlc"
version = "0.1.0"
description = "Queue-orchestrated, agent-driven SDLC platform. Thin harness, fat skills."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # LLM providers
    "anthropic>=0.39.0",
    "openai>=1.50.0",

    # API + async
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "websockets>=13.0",

    # Persistence
    "redis>=5.0.0",
    "neo4j>=5.23.0",
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",

    # Data validation + config
    "pydantic>=2.7.0",
    "pyyaml>=6.0.0",
    "python-frontmatter>=1.1.0",     # skill markdown + YAML frontmatter

    # Secrets
    "hvac>=2.3.0",
    "python-dotenv>=1.0.0",

    # Auth (GitHub App)
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
